README.md
A personal portfolio website is a personalized online platform that presents an individual's skills, achievements, and projects in a visually engaging and easily accessible manner. It serves as a digital resume, allowing individuals to showcase their work, expertise, and professional growth to potential employers, clients, or collaborators. It provides a centralized hub for displaying and highlighting one's talents and accomplishments.